requirements.txt
Flask
Gunicorn
requests
schedule
bs4
python-dotenv
